+++ src/tepkit/io/vasp/__init__.py
@@ -1,4 +1,4 @@
 from ._eigenval import Eigenval
-from ._kpoints import ExplicitKpoints, Ibzkpt, Kpoints
+from ._kpoints import ExplicitKpoints, Ibzkpt, Kpoints, RegularKpoints
 from ._outcar import Outcar
 from ._poscar import Poscar, VaspCoordinatesMode